secret.py
from google.cloud import secretmanager

# project_id = "test-python-api-spreadsheets"
# project_id = "blockmacro-7b611"


def access_secret(secret_id, project_id, version_id="latest"):
    # Create the Secret Manager client.
    client = secretmanager.SecretManagerServiceClient()
    # Build the resource name of the secret version.
    name = f"projects/{project_id}/secrets/{secret_id}/versions/{version_id}"
    # Access the secret version.
    response = client.access_secret_version(name=name)
    # Return the decoded payload.
    return response.payload.data.decode('UTF-8')